src/main/java/br/com/jdp/linkedlist/DetectCycleLinkedList2.java
package br.com.jdp.linkedlist;

// https://leetcode.com/problems/linked-list-cycle-ii
public class DetectCycleLinkedList2 {

    public static ListNode detectCycle(ListNode head) {
        ListNode slow = head;
        ListNode fast = head;

        // Hare-Tortoise algorithm approach
        while (fast != null && fast.next != null){
            slow = slow.next;
            fast = fast.next.next;

            if (slow == fast) {
                slow = head;
                while (slow != fast) {
                    slow = slow.next;
                    fast = fast.next;
                }
                return slow;
            }
        }
        return null;
    }
}
